requirements.txt
ragflow
openai
sentence-transformers==2.6.1
faiss-cpu
python-dotenv
pyyaml
torch>=1.11.0
numpy<2
tqdm
scikit-learn
scipy
transformers<5.0.0,>=4.32.0
huggingface-hub>=0.15.1
pillow
requests
filelock
packaging
regex
tokenizers<0.22,>=0.21
safetensors
fsspec
typing-extensions
hf-xet
sympy
networkx
jinja2
MarkupSafe
joblib
threadpoolctl
mpmath